Fixed contradicting convar "sv_vote_issue_pause_match_spec_only
The convar "sv_vote_issue_pause_match_spec_only" is listed twice. Once in the list of convars that must have an exact value and once under the category with more freedom.

Based on the in-game description of this cvar - "When enabled, only admins start technical pause" and given that it should be set to 1, I assume this was done to prevent some sort of abuse by the players hence why removed it from the other list.

Feel free to make the change the other way around by leaving it in the loose convar list and removing it from the strict one.